refactor InMemoryRepository with new FileOperations service + refactor tests
